refactor: isolate run_skill completion tools
