+++ tests/integration/test_pipeline_step_completion_flow.py
@@ -16,10 +16,10 @@ from mcp.types import CallToolRequestParams, TextContent
 from autoskillit.core.types import RetryReason
 from autoskillit.core.types._type_results import SkillResult
 from autoskillit.server._run_skill_completion import RunSkillCompletionMiddleware
-from autoskillit.server.tools.tools_execution import (
+from autoskillit.server.tools.tools_execution import run_skill
+from autoskillit.server.tools.tools_run_skill_completion import (
     complete_run_skill_result,
     recover_run_skill_result,
-    run_skill,
 )
 from tests.server._pipeline_test_helpers import (
     _ack_direct_run_skill_result,
